pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ep-mcp"
version = "0.1.0"
description = "ExpertPack MCP Server — expertise-as-a-service over MCP"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Brian Hearn", email = "brian@easyterritory.com" },
]

dependencies = [
    "mcp[cli]>=1.0",
    "uvicorn>=0.30",
    "starlette>=0.40",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "sqlite-vec>=0.1",
    "google-genai>=0.8",
    "openai>=1.50",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
ep-mcp = "ep_mcp.cli:cli"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
